fix(rust): make `cargo clippy --all-targets -- -D warnings` clean
Removing the crate-wide #![allow(dead_code)] exposed device-only FFI items
that are dead in host/test builds, plus a few test-code nits. Gate each to
the config where it's actually used instead of re-silencing:

- gate the device-only externs (check_canary, app_mode_expert, hash_sha256,
  get_token) and the CTokenInfo/CONTRACT_ADDR_STR_MAX_LEN they use behind
  cfg(not(any(test, feature = "fuzzing")))
- move the #[panic_handler] behind cfg(target_os = "none") so host clippy
  builds no longer clash with std's panic_impl (fixes E0152)
- drop the unused with_other test helper and redundant `as u8` cast, and
  rename the inValue/outValue test locals to snake_case

+++ app/rust/src/parser/tx_post_conditions.rs
@@ -260,10 +260,7 @@ mod test {
         let a = nft_cond(0x03, 0x12);
         let b = nft_cond(0x04, 0x12);
         let conds = [a.as_slice(), b.as_slice()];
-        assert_eq!(
-            PostConditions::get_num_items(&conds),
-            AGGREGATED_NFT_ITEMS as u8
-        );
+        assert_eq!(PostConditions::get_num_items(&conds), AGGREGATED_NFT_ITEMS);
     }
 
     #[test]
